# src/main/resources/static/swagger/swagger-v1.yaml
openapi: 3.0.0
info:
  title: Partner Api
  description: Open Api Specification
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: local

externalDocs:
  url: https://github.com/FernandoTAA/partner/blob/main/README.md
  description: For more information about this api

paths:
  /partner:
    post:
      tags:
        - Partner
      summary: Creates a partner
      description: >
        This endpoint creates a partner
      operationId: postPartnerBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerBatchRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          description: Internal Server Error
      externalDocs:
        description: Learn more about partner operation provided by this API
        url: https://github.com/FernandoTAA/partner/wiki/%5BAPI%5D-POST-partner-(v1)
    get:
      tags:
        - Partner
      summary: Get a partner by point
      description: >
        This endpoint get a partner by point
      operationId: getByPoint
      parameters:
        - name: longitude
          in: query
          required: true
          schema:
            type: number
        - name: latitude
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerResponse'
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Internal Server Error
      externalDocs:
        description: Learn more about partner operation provided by this API
        url: https://github.com/FernandoTAA/partner/wiki/%5BAPI%5D-GET-partner-by-id-(v1)
  /partner/{id}/:
    get:
      tags:
        - Partner
      summary: Get a partner by id
      description: >
        This endpoint get a partner by id
      operationId: getPartnerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Internal Server Error
      externalDocs:
        description: Learn more about partner operation provided by this API
        url: https://github.com/FernandoTAA/partner/wiki/%5BAPI%5D-GET-partner-by-id-(v1)

components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorSimple'

  schemas:
    PartnerRequest:
      type: object
      properties:
        tradingName:
          type: string
          example: Adega Pinheiros
        ownerName:
          type: string
          example: Ze da Silva
        document:
          type: string
          example: 4.433.714/0001-44
      required:
        - tradingName
        - ownerName
        - document

    PartnerBatchRequest:
      type: object
      properties:
        pdvs:
          type: array
          items:
            $ref: '#/components/schemas/PartnerRequest'
      required:
        - pdvs

    PartnerBatchResponse:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
            example: 6024991257427b090579566a

    PartnerResponse:
      type: object
      properties:
        id:
          type: string
          example: 6024991257427b090579566a
        tradingName:
          type: string
          example: Adega Pinheiros
        ownerName:
          type: string
          example: Ze da Silva
        document:
          type: string
          example: 4.433.714/0001-44



    ApiError:
      type: object
      properties:
        message:
          type: string
          example: error message
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'

    ApiErrorSimple:
      type: object
      properties:
        message:
          type: string
          example: error message

    Error:
      type: object
      properties:
        scope:
          type: string
          example: path
        field:
          type: string
          example: field_name
        value:
          type: string
          example: field_value
        message:
          type: string
          example: this field contains error